requirements.txt
streamlit
moviepy
webvtt-py 